Refactor emails list into Emails
